v1.2.0正式版本
v1.2.0版本，该版本主要对后台界面做了调整，同时扩展了栏目和文章的数据库字段，栏目可以设置栏目类型，可以拥有栏目自己的模板，模板采用新的正则表达式模板引擎，

方便编写模板，添加分页，搜索功能，默认添加了几个栏目类型，如关于我们，站点公告，友情链接，外链，这些都是需要配合mydefined主题风格才能正常显示的，所以默认

使用的也是mydefined主题，对应tpl/mydefined目录，该目录显示了哪些模板是用于前台显示的，可以对应复制一个mydefined目录，将其改为自己的名字，生成自己的主题风格，

再在后台系统配置中修改主题即可。添加了水印功能，文章编辑时，有自动保存为草稿功能。还添加了些其他的方便后台操作的功能。数据库备份恢复采用批处理事务方式。
作者：zenglong
时间：2013年7月6日

+++ common_fun/help_func.php
@@ -136,6 +136,8 @@ function get_jmp_locs($cur)
 				echo '<a href="'.$zengl_cms_rootdir .$arg[0].'">'. $arg[1] .'</a>&nbsp;&nbsp;';
 		}
 	}
+	else
+		echo '<a href="javascript:history.go(-1);">[返回]</a>&nbsp;&nbsp;';
 }
 
 function subUTF8($string, $length = 80, $etc = '...')
@@ -349,30 +351,64 @@ function mkdirs($dir)
 	}
 	return true;
 }
-function rmdirs($dir)
+function rmdirs($dir,$progress = null)
 {
 	$d = dir($dir);
+	if(!is_dir($dir))
+		return ;
 	while (false !== ($child = $d->read()))
 	{
 		if($child != '.' && $child != '..')
 		{
 			if(is_dir($dir.'/'.$child))  
 			{
-				rmdirs($dir.'/'.$child);  
-				echo '删除目录：'.$dir.'/'.$child.'<br/>';
-				flush_buffers();
+				rmdirs($dir.'/'.$child,$progress);
 			}
 			else 
 			{
 				unlink($dir.'/'.$child);
-				echo '删除文件：'.$dir.'/'.$child.'<br/>';
-				flush_buffers();
+				if($progress !== null)
+					$progress->step('删除文件：'.$dir.'/'.$child);
+				else
+				{
+					echo '删除文件：'.$dir.'/'.$child.'<br/>';
+					flush_buffers();
+				}
 			}
 		}
 	}  
-	$d->close();  
-	rmdir($dir);
-	echo '删除目录：'.$dir.'<br/>';
-	flush_buffers();
+	$d->close();
+	if(is_dir($dir))
+	{
+		rmdir($dir);
+		if($progress !== null)
+			$progress->step('删除目录：'.$dir);
+		else 
+		{
+			echo '删除目录：'.$dir.'<br/>';
+			flush_buffers();
+		}
+	}
+}
+
+function getdirs($dir)
+{
+	$count = 0;
+	if(!is_dir($dir))
+		return $count;
+	$d = dir($dir);
+	while (false !== ($child = $d->read()))
+	{
+		if($child != '.' && $child != '..')
+		{
+			if(is_dir($dir.'/'.$child))
+				$count += getdirs($dir.'/'.$child);
+			else
+				$count++;
+		}
+	}
+	if(is_dir($dir))
+		$count++;
+	return $count;
 }
 ?>
